requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
duckduckgo-search==4.1.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
markdown==3.5.1
python-dotenv==1.0.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
